# packages/contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "MediaHub API",
    "version": "0.4.0"
  },
  "paths": {
...
    "/api/v1/auth/login": {
      "post": {
        "summary": "Login",
        "operationId": "login_api_v1_auth_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Credentials"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/cloudflare/status": {
      "get": {
        "summary": "Cloudflare Status",
        "operationId": "cloudflare_status_api_v1_cloudflare_status_get",
        "parameters": [
          {
            "name": "refresh",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Refresh"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/apps/{app_id}": {
      "get": {
        "summary": "App Detail",
        "operationId": "app_detail_api_v1_apps__app_id__get",
        "parameters": [
          {
            "name": "app_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "App Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/apps/{app_id}/runtime": {
      "get": {
        "summary": "App Runtime",
        "operationId": "app_runtime_api_v1_apps__app_id__runtime_get",
        "parameters": [
          {
            "name": "app_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "App Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/apps/{app_id}/actions/{action}": {
      "post": {
        "summary": "App Action",
        "operationId": "app_action_api_v1_apps__app_id__actions__action__post",
        "parameters": [
          {
            "name": "app_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "App Id"
            }
          },
          {
            "name": "action",
            "in": "path",
            "required": true,
            "schema": {
              "enum": [
                "start",
                "stop",
                "restart",
                "restart-vpn",
                "restart-qbittorrent",
                "test-vpn"
              ],
              "type": "string",
              "title": "Action"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/apps/{app_id}/logs": {
      "get": {
        "summary": "App Logs",
        "operationId": "app_logs_api_v1_apps__app_id__logs_get",
        "parameters": [
          {
            "name": "app_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "App Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/apps/{app_id}/update-check": {
      "post": {
        "summary": "App Update Check",
        "operationId": "app_update_check_api_v1_apps__app_id__update_check_post",
        "parameters": [
          {
            "name": "app_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "App Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/apps/{app_id}/uninstall": {
      "post": {
        "summary": "Remove Runtime",
        "operationId": "remove_runtime_api_v1_apps__app_id__uninstall_post",
        "parameters": [
          {
            "name": "app_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "App Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RemoveRuntimeRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/storage": {
      "get": {
        "summary": "Storage",
        "operationId": "storage_api_v1_storage_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      },
      "post": {
        "summary": "Storage Register",
        "operationId": "storage_register_api_v1_storage_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StorageInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/storage/validate": {
      "post": {
        "summary": "Storage Validate",
        "operationId": "storage_validate_api_v1_storage_validate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PathRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/events/history": {
      "get": {
        "summary": "Activity",
        "operationId": "activity_api_v1_events_history_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "maximum": 100,
                  "minimum": 1
                },
                {
                  "type": "null"
                }
              ],
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/settings": {
      "get": {
        "summary": "Settings",
        "operationId": "settings_api_v1_settings_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      },
      "put": {
        "summary": "Save Settings",
        "operationId": "save_settings_api_v1_settings_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PlatformSettings"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/notifications/{notification_id}/read": {
      "post": {
        "summary": "Read Notification",
        "operationId": "read_notification_api_v1_notifications__notification_id__read_post",
        "parameters": [
          {
            "name": "notification_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Notification Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/updates/platform/credentials": {
      "get": {
        "summary": "Platform Update Credentials",
        "operationId": "platform_update_credentials_api_v1_updates_platform_credentials_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      },
      "put": {
        "summary": "Save Platform Update Credentials",
        "operationId": "save_platform_update_credentials_api_v1_updates_platform_credentials_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GitHubTokenRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Remove Platform Update Credentials",
        "operationId": "remove_platform_update_credentials_api_v1_updates_platform_credentials_delete",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Credentials": {
        "properties": {
          "username": {
            "type": "string",
            "maxLength": 80,
            "minLength": 1,
            "title": "Username"
          },
          "password": {
            "type": "string",
            "maxLength": 256,
            "minLength": 1,
            "title": "Password"
          },
          "secondFactor": {
            "type": "string",
            "maxLength": 32,
            "title": "Secondfactor",
            "default": ""
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "username",
          "password"
        ],
        "title": "Credentials"
      },
...
      "GitHubTokenRequest": {
        "properties": {
          "token": {
            "type": "string",
            "maxLength": 512,
            "minLength": 20,
            "title": "Token"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "token"
        ],
        "title": "GitHubTokenRequest"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PathRequest": {
        "properties": {
          "path": {
            "type": "string",
            "maxLength": 4096,
            "minLength": 1,
            "title": "Path"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "path"
        ],
        "title": "PathRequest"
      },
...
      "PlatformSettings": {
        "properties": {
          "display_name": {
            "type": "string",
            "maxLength": 60,
            "minLength": 1,
            "title": "Display Name",
            "default": "MediaHub"
          },
          "theme": {
            "type": "string",
            "enum": [
              "dark",
              "light",
              "system"
            ],
            "title": "Theme",
            "default": "dark"
          },
          "activity_page_size": {
            "type": "integer",
            "maximum": 100.0,
            "minimum": 10.0,
            "title": "Activity Page Size",
            "default": 50
          },
          "advanced_mode": {
            "type": "boolean",
            "title": "Advanced Mode",
            "default": false
          },
          "release_repository": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 201
              },
              {
                "type": "null"
              }
            ],
            "title": "Release Repository"
          },
          "update_check_interval_hours": {
            "type": "integer",
            "enum": [
              0,
              1,
              6,
              12,
              24,
              72,
              168
            ],
            "title": "Update Check Interval Hours",
            "default": 24
          },
          "visible_navigation": {
            "items": {
              "type": "string",
              "enum": [
                "/",
                "/apps",
                "/storage",
                "/hosts",
                "/activity",
                "/logs",
                "/updates",
                "/backups",
                "/integrations",
                "/settings"
              ]
            },
            "type": "array",
            "title": "Visible Navigation",
            "default": [
              "/",
              "/apps",
              "/storage",
              "/updates",
              "/backups",
              "/settings"
            ]
          },
          "dashboard_sections": {
            "items": {
              "type": "string",
              "enum": [
                "system",
                "storage",
                "apps",
                "activity",
                "network",
                "core",
                "runtime",
                "integrations",
                "cloudflare"
              ]
            },
            "type": "array",
            "title": "Dashboard Sections",
            "default": [
              "storage",
              "apps",
              "system"
            ]
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "PlatformSettings"
      },
...
      "RemoveRuntimeRequest": {
        "properties": {
          "confirmedInstallationId": {
            "type": "string",
            "maxLength": 80,
            "minLength": 1,
            "title": "Confirmedinstallationid"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "confirmedInstallationId"
        ],
        "title": "RemoveRuntimeRequest"
      },
...
      "StorageInput": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 80,
            "minLength": 1,
            "title": "Name"
          },
          "kind": {
            "type": "string",
            "enum": [
              "downloads",
              "incomplete_downloads",
              "completed_downloads",
              "movies",
              "tv",
              "appdata",
              "backups",
              "temp",
              "custom"
            ],
            "title": "Kind"
          },
          "path": {
            "type": "string",
            "maxLength": 4096,
            "minLength": 1,
            "title": "Path"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "name",
          "kind",
          "path"
        ],
        "title": "StorageInput"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}